Move StylesheetSetRef to script
