+++ components/script/lib.rs
@@ -69,6 +69,7 @@ pub mod script_thread;
 mod serviceworker_manager;
 mod serviceworkerjob;
 mod stylesheet_loader;
+mod stylesheet_set;
 mod task_manager;
 mod task_queue;
 mod task_source;
